+++ vizier/backend/test/src/info/vizierdb/catalog/PublishArtifactSpec.scala
@@ -76,78 +76,4 @@ class PublishArtifactSpec extends Specification with BeforeAll
 
     project.artifacts.keys must contain(IMPORT_NAME)
   }
-
-  "testing profiler" >> {
-    val project = MutableProject("test_profiler")
-    project.load(
-      file = "test_data/output_with_nulls.csv",
-      name = "test", 
-      format = "csv",
-    )
-
-    GetArtifact(
-      projectId = project.projectId,
-      artifactId = project.artifact("test").id,
-      profile = Some("true"),
-    )
-
-    val df = project.dataframe("test")
-    val art = project.artifact("test")
-    val properties = art.datasetDescriptor.properties
-
-    properties.get("is_profiled") must beSome.which(_ == JsBoolean(true))
-  }
-
-    "testing correct profiler information" >> {
-    val project = MutableProject("test_profiler")
-    project.load(
-      file = "test_data/output_with_nulls.csv",
-      name = "test", 
-      format = "csv",
-    )
-
-    GetArtifact(
-      projectId = project.projectId,
-      artifactId = project.artifact("test").id,
-      profile = Some("true"),
-    )
-
-    val df = project.dataframe("test")
-    val art = project.artifact("test")
-    val properties = art.datasetDescriptor.properties
-    val columnsProperty = properties.get("columns")
-
-    columnsProperty match {
-      case Some(jsValue: JsValue) =>
-        // Parse the JSON string
-        val json = Json.parse(jsValue.toString())
-
-        // Validate and extract columns array
-        (json \ "columns").validate[JsArray] match {
-          case JsSuccess(columnsArray, _) =>
-            columnsArray.value.foreach { column =>
-              // Extract and test distinctValueCount
-              val distinctValueCount = (column \ "distinctValueCount").as[Int]
-              distinctValueCount must beEqualTo(0)
-
-              // Extract and test nullCount
-              val nullCount = (column \ "nullCount").as[Int]
-              nullCount must beEqualTo(25)
-
-              println(s"Column Test Passed: distinctValueCount: $distinctValueCount, nullCount: $nullCount")
-            }
-
-          case JsError(errors) =>
-            failure(s"Error parsing columns JSON: $errors")
-        }
-
-      case _ => 
-        failure("Columns property not found or not a valid JSON object")
-    }
-
-    ok
-  }
-
-
-
 }
